refactor: move mobile-action-page in adaptor-of-desktop folder
